requirements.txt
# Core API
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9

# Cache
redis>=5.0.1
hiredis>=2.2.3

# Tasks
celery>=5.3.4
celery[redis]>=5.3.4

# HTTP Requests
requests>=2.31.0

# Data Processing
pandas>=2.2.0

# Rate Limiting & Security
slowapi>=0.1.9

# Chatbot (opcional)
langchain>=0.0.350
langchain-openai>=0.0.2
openai>=1.6.0
tiktoken>=0.5.2

# Monitoring
prometheus-client>=0.19.0
sentry-sdk>=1.38.0

# WebSocket (opcional)
websockets>=12.0
python-socketio>=5.10.0

# Git (mantido do código original)
GitPython>=3.1.40

# Security
python-multipart>=0.0.6

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
